Cleanup importer
