+++ src/Assimp/CMakeLists.txt
@@ -1,0 +1,15 @@
+##############################################################
+#  adding the files
+##############################################################
+
+set(SRC_FILES 
+${SRC_FILES}
+${CMAKE_CURRENT_SOURCE_DIR}/Importer.cpp
+PARENT_SCOPE
+)
+
+set(HEADER_FILES
+${HEADER_FILES}
+${CMAKE_CURRENT_SOURCE_DIR}/Importer.h
+PARENT_SCOPE
+)
