+++ src/CMakeLists.txt
@@ -69,10 +69,10 @@ ObjectItemModel.h
 MaterialWidget.h
 MaterialComboDelegate.h
 MaterialHighlighter.h
-AssimpToOgreImporter.h
 )
 
 ADD_SUBDIRECTORY("${CMAKE_CURRENT_SOURCE_DIR}/OgreXML")
+ADD_SUBDIRECTORY("${CMAKE_CURRENT_SOURCE_DIR}/Assimp")
 
 #file(GLOB UI_FILES ./ui_files/*.ui)
 # if we don't include this CMake will not include ui headers properly:
